chore(964): unify UI empty page
