+++ lib/pages/user_action/user_action_list_page.dart
@@ -22,6 +22,7 @@ import 'package:pass_emploi_app/widgets/buttons/primary_action_button.dart';
 import 'package:pass_emploi_app/widgets/cards/campagne_card.dart';
 import 'package:pass_emploi_app/widgets/cards/user_action_card.dart';
 import 'package:pass_emploi_app/widgets/default_animated_switcher.dart';
+import 'package:pass_emploi_app/widgets/empty_page.dart';
 import 'package:pass_emploi_app/widgets/loader.dart';
 import 'package:pass_emploi_app/widgets/retry.dart';
 
@@ -86,11 +87,9 @@ class _UserActionListPageState extends State<UserActionListPage> {
   Widget _animatedBody(BuildContext context, UserActionListPageViewModel viewModel) {
     if (viewModel.withLoading) return loader();
     if (viewModel.withFailure) return Center(child: Retry(Strings.actionsError, () => viewModel.onRetry()));
-    if (viewModel.withEmptyMessage) return _empty();
+    if (viewModel.withEmptyMessage) return Empty(description: Strings.noActionsYet);
     return _userActionsList(context, viewModel);
   }
-
-  Widget _empty() => Center(child: Text(Strings.noActionsYet, style: TextStyles.textSmRegular()));
 
   Widget _userActionsList(BuildContext context, UserActionListPageViewModel viewModel) {
     return ListView.separated(
